Land the round-2 spot-check findings
From two fresh reader passes over 16 files (matchdiff MATCH each,
whole image byte-identical):
- The status7_* result/label family in the four Attack* controllers is
  attack_* — the raw status number stopped meaning anything once
  STAT_ATTACK existed. vmemoryGC's search3/found3/sz2 lose the
  leftover draft suffixes and FreePoolBlockInline is free_block.
- AttackIndirect: the declaration-less bare block unwrapped, both
  20000s named INDIRECT_RANGE (tuning.h, beside SR_CLEAR_RANGE).
- UpdateEvent: !(x < 2) is x >= 2; the offset-walk address stays and
  now carries its measured index-first-addu adjudication.
- BreedLife: the kind temp inlined away (matched), the wrong ninja-dog
  "leash flag" header claim corrected to the starter kusuri the code
  gives, and the stale plain-if/else note restated for the goto ladder.
- Briefing: d-pad mask spelled with PADL names, 0x20/0x40 are
  PADRright/PADRdown via pad.s (the union's own signed view), the
  (int)j7 cast and j = shown loop seed dropped (matched), radix
  mismatches fixed, MAX_SELECTED_ITEMS 6 named; the p alias over spr
  proved byte-required and is adjudicated.
- Think3firstattack compares against WPATK_CLASS_RANGED;
  debug_menu_stage_option uses the LANG_* names; AttackBowControl's
  no-op (s16)n cast dropped.

+++ src/main.exe/AttackBowControl.c
@@ -125,7 +125,7 @@ void AttackBowControl(s16 n)
     }
     else
     {
-        idx = (s16)n << 2;
+        idx = n << 2;
         p = (u8 *)BowTiming + idx;
         if (*(s16 *)p <= count && count < *(s16 *)(p + 2))
         {
